+++ packages/bpk-component-calendar/index.js
@@ -1,5 +1,6 @@
 import BpkCalendar from './src/BpkCalendar';
 import BpkCalendarGrid from './src/BpkCalendarGrid';
+import BpkCalendarGridHeader from './src/BpkCalendarGridHeader';
 import BpkCalendarNav from './src/BpkCalendarNav';
 import BpkCalendarDate from './src/BpkCalendarDate';
 import BpkCalendarView from './src/BpkCalendarView';
@@ -9,6 +10,7 @@ import * as DateUtils from './src/date-utils';
 export default BpkCalendar;
 export {
   BpkCalendarGrid,
+  BpkCalendarGridHeader,
   BpkCalendarNav,
   BpkCalendarDate,
   BpkCalendarView,